Modularize code (#28)
* v1.1.0

* Date assesor should be a function

* Make display event its own file

* Crawl facebook is its own file

* Set up wizard modularization

* Update debug file

* Cleaning up before deployment

* save

* Update src/appSettings.js


* Update src/crawlFacebook.js


* Update src/initializer.js


* Quick fix

* Update src/crawlFacebook.js


* Close app if pref file can't be read

* Update src/dateAssesor.js


* Use template literal rather than weird ass strings

* Display errors when user fucks up

* Ignore lgtm issue

* Save pages in the same file throughout the app

* Disable one eslint rule

* Restyled by prettier (#37)


* Revert "Restyled by prettier (#37)"

This reverts commit 19efba40a78351254c2fb627e9499c116d33a1e2.

* Remove restyled

* Fix bug in date evaluation

* Allow LGTM to stop freaking out

* Display a message if the page has no events

* Get ready for new modularization

* Fix circular dependancy in addPages

* Modularize code whilst avoiding loops
