+++ .gitignore
@@ -3,6 +3,7 @@ screenshot
 screenshots
 test.csv
 pages.csv
+graph.svg
 
 # Logs
 logs
