+++ package.json
@@ -1,9 +1,9 @@
 {
 	"name": "facebook_event_tracker",
-	"version": "1.0.1",
+	"version": "1.1.0",
 	"description": "A CLI tool that enables you to track a FB Page's events without needing a Facebook account",
 	"author": "Adrien D. Ahlqvist",
-	"bugs": "https://github.com/Armitage35/fuzzy-roadmap/issues",
+	"bugs": "https://github.com/Armitage35/facebook-event-tracker/issues",
 	"repository": {
 		"type": "git",
 		"url": "github.com:Armitage35/facebook-event-tracker.git"
