+++ src/debug.csv
@@ -1,3 +1,4 @@
+Individual pages to test
 https://www.facebook.com/pg/0gravite/events/
 https://www.facebook.com/pg/cafeblocmtl/events/
 https://www.facebook.com/pg/BlocShop/events/
@@ -5,3 +6,6 @@ https://www.facebook.com/pg/allezup/events/
 https://www.facebook.com/pg/ShaktiRockGym/events/
 https://www.facebook.com/pg/LeMouvEspaceBloc/events/
 https://www.facebook.com/pg/betabloc/events/
+
+Import batch
+https://www.facebook.com/pg/0gravite/events/,https://www.facebook.com/pg/cafeblocmtl/events/,https://www.facebook.com/pg/BlocShop/events/,https://www.facebook.com/pg/allezup/events/,https://www.facebook.com/pg/ShaktiRockGym/events/,https://www.facebook.com/pg/LeMouvEspaceBloc/events/,https://www.facebook.com/pg/betabloc/events/
