+++ src/strings.json
@@ -46,7 +46,8 @@
 		},
 		"success": {
 			"crawlingCompleted": "Crawling completed 👍",
-			"crawlingStarted": "Crawling launched 🚀"
+			"crawlingStarted": "Crawling launched 🚀",
+			"noPageToDisplay": "This page dooes not have any events for now 😕"
 		},
 		"onboarding": "✋ Hello user and thank you for using this application. First, let's find out which pages you want to know about"
 	}
